gsm_04_08, gsm_subscriber: decouple lac from bts
The idea is to not have a direct pointer to a bts struct (into BSC land), but a
LAC to resolve the BSC or RNC depending on the appropriate A or IuCS interface.

subscr_update(): remove bts arg, add lac arg.

Pass conn->lac to gsm48_generate_lai() instead of bts->location_area_code.

Change-Id: I9f2b298a785bf4b2a1b3fcdd91b8256106b2d9de